requirements.txt
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.47.0
python-dotenv==0.19.2
requests==2.26.0